refactor: split database facade operations
